raftstore: move some impls around
